// frontend/src/lib/api/generated/models/assignment.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface Assignment {
  account: string;
  github_user_id: number;
  host_id: string;
  maintenance: boolean;
  name: string;
  node_id: string;
  protocol: number;
  role: string;
  ssh_address: string;
  ssh_host_fingerprint: string;
  /** @minimum 0 */
  uid: number;
  url: string;
}