// frontend/src/lib/api/generated/models/connection.ts
/**
 * Generated by Orval. Do not edit manually.
 */

export interface Connection {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  account: string;
  github_user_id: number;
  host_id: string;
  id: string;
  maintenance: boolean;
  name: string;
  node_id: string;
  protocol: number;
  registry_id: string;
  revision: string;
  role: string;
  ssh_address: string;
  ssh_host_fingerprint: string;
  /** @minimum 0 */
  uid: number;
  url: string;
}